ref: extract types and utils to own file
